refactor: changed folder structure by adding folder for Chats, Header, Dashboard and SideNAv component
