+++ src/App.tsx
@@ -1,10 +1,10 @@
 import { Route, Routes } from "react-router-dom";
 import "./App.css";
 import "bootstrap-icons/font/bootstrap-icons.css";
-import Dashboard from "./components/Dashboard";
+import Dashboard from "./components/Dashboard/Dashboard";
 import BrandVoice from "./components/BrandVoice/BrandVoice";
 import BrandTone from "./components/BrandTone/BrandTone";
-import Chat from "./components/Chat";
+import Chat from "./components/Chats/Chat";
 import { useState } from "react";
 
 function App() {
